Refactor : Refactor class architecture
* model 하위에 있던 클래스들을 http.model 위치로 변경.
