+++ src/main/java/webserver/http/model/PathInformation.java
@@ -1,4 +1,4 @@
-package model.http;
+package webserver.http.model;
 
 import exception.IllegalHttpRequestException;
