+++ src/main/java/webserver/http/model/Protocol.java
@@ -1,4 +1,4 @@
-package model.http;
+package webserver.http.model;
 
 import java.util.Objects;
