+++ src/main/java/webserver/http/model/ProtocolVersion.java
@@ -1,4 +1,4 @@
-package model.http;
+package webserver.http.model;
 
 import exception.IllegalHttpRequestException;
