+++ src/main/java/webserver/http/model/QueryStrings.java
@@ -1,4 +1,4 @@
-package model.http;
+package webserver.http.model;
 
 import java.util.ArrayList;
 import java.util.List;
