+++ src/main/java/webserver/http/model/RequestBody.java
@@ -1,4 +1,4 @@
-package model.http;
+package webserver.http.model;
 
 import java.util.Arrays;
 import java.util.Map;
