+++ src/test/java/webserver/http/model/RequestHeaderTest.java
@@ -1,4 +1,4 @@
-package model.http;
+package webserver.http.model;
 
 import org.junit.jupiter.api.Test;
