+++ src/test/java/webserver/http/model/RequestLineTest.java
@@ -1,4 +1,4 @@
-package model.http;
+package webserver.http.model;
 
 import exception.IllegalHttpRequestException;
 import org.junit.jupiter.api.DisplayName;
